test-02 changes (#6)
* Updated GH urls to SnappyData 1.0.0 and Apache Zeppelin 0.7.3 artifacts.
* Point the READMEs to main product documentation.
* Initial support for SnappyData Enterprise AMIs (WIP)
* Changed default instance types.
* Now Apache Zeppelin server can be started only in the lead node instance itself.
* Updated license headers.

+++ aws/ec2/deploy/home/ec2-user/snappydata/aws-setup.sh
@@ -1,25 +1,36 @@
 #!/bin/bash
-
 #
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
+# Copyright (c) 2017 SnappyData, Inc. All rights reserved.
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+# Licensed under the Apache License, Version 2.0 (the "License"); you
+# may not use this file except in compliance with the License. You
+# may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+# implied. See the License for the specific language governing
+# permissions and limitations under the License. See accompanying
+# LICENSE file.
 #
 
 pushd /home/ec2-user/snappydata > /dev/null
 
 source ec2-variables.sh
+
+# Check if enterprise version to be used.
+if [[ "${SNAPPYDATA_VERSION}" = "ENT" ]]; then
+  echo "Setting up the cluster with SnappyData Enterprise edition ..."
+  # wget -q "https://raw.githubusercontent.com/SnappyDataInc/snappy-cloud-tools/test-02/aws/ec2/deploy/home/ec2-user/snappydata/ent-aws-setup.sh"
+  sh ent-aws-setup.sh
+  ENT_SETUP=`echo $?`
+  # rm -f ent-aws-setup.sh
+  popd > /dev/null
+  exit ${ENT_SETUP}
+fi
+
 sudo yum -y -q remove  jre-1.7.0-openjdk
 sudo yum -y -q install java-1.8.0-openjdk-devel
 
@@ -48,10 +59,10 @@ else
   cp locator_list "${SNAPPY_HOME_DIR}/conf/locators"
 fi
 
-# Enable jmx-manager for pulse to start
-sed -i '/^#/ ! {/\\$/ ! { /^[[:space:]]*$/ ! s/$/ -jmx-manager=true -jmx-manager-start=true/}}' "${SNAPPY_HOME_DIR}/conf/locators"
+# Enable jmx-manager for pulse to start - DISCONTINUED with SnappyData 0.9
+# sed -i '/^#/ ! {/\\$/ ! { /^[[:space:]]*$/ ! s/$/ -jmx-manager=true -jmx-manager-start=true/}}' "${SNAPPY_HOME_DIR}/conf/locators"
 # Configure hostname-for-clients
-sed -i '/^#/ ! {/\\$/ ! { /^[[:space:]]*$/ ! s/\([^ ]*\)\(.*\)$/\1\2 -J-Dgemfirexd.hostname-for-clients=\1/}}' "${SNAPPY_HOME_DIR}/conf/locators"
+sed -i '/^#/ ! {/\\$/ ! { /^[[:space:]]*$/ ! s/\([^ ]*\)\(.*\)$/\1\2 -hostname-for-clients=\1/}}' "${SNAPPY_HOME_DIR}/conf/locators"
 
 
 if [[ -e leads ]]; then
@@ -60,16 +71,19 @@ else
   cp lead_list "${SNAPPY_HOME_DIR}/conf/leads"
 fi
 
-if [[ "${ZEPPELIN_HOST}" != "zeppelin_server" ]]; then
-  # Enable interpreter on lead
-  sed -i '/^#/ ! {/\\$/ ! { /^[[:space:]]*$/ ! s/$/ -zeppelin.interpreter.enable=true/}}' "${SNAPPY_HOME_DIR}/conf/leads"
+INTERPRETER_VERSION="0.7.2"
 
+if [[ "${ZEPPELIN_HOST}" != "NONE" ]]; then
   # Add interpreter jar to snappydata's jars directory
-  # TODO Download this from official-github-release. See fetch-distribution.sh:getLatestUrl() on how we can get the latest url.
-  INTERPRETER_JAR="snappydata-zeppelin-0.7.1.jar"
-  INTERPRETER_URL="https://github.com/SnappyDataInc/zeppelin-interpreter/releases/download/v0.7.1/${INTERPRETER_JAR}"
+  INTERPRETER_JAR="snappydata-zeppelin-${INTERPRETER_VERSION}.jar"
+  INTERPRETER_URL="https://github.com/SnappyDataInc/zeppelin-interpreter/releases/download/v${INTERPRETER_VERSION}/${INTERPRETER_JAR}"
   wget -q "${INTERPRETER_URL}"
-  mv ${INTERPRETER_JAR} ${SNAPPY_HOME_DIR}/jars/
+  mv ${INTERPRETER_JAR} ${SNAPPY_HOME_DIR}/snappydata-zeppelin.jar
+  ZEPPELIN_JAR_PATH=`readlink -f ${SNAPPY_HOME_DIR}/snappydata-zeppelin.jar`
+  ZEPPELIN_JAR_PATH=`echo "$ZEPPELIN_JAR_PATH"|sed 's@/$@@'`
+
+  # Enable interpreter on lead
+  sed -i "/^#/ ! {/\\$/ ! { /^[[:space:]]*$/ ! s/$/ -zeppelin.interpreter.enable=true -classpath=${ZEPPELIN_JAR_PATH}/}}" "${SNAPPY_HOME_DIR}/conf/leads"
 fi
 
 if [[ -e servers ]]; then
@@ -79,7 +93,7 @@ else
 fi
 
 # Configure hostname-for-clients
-sed -i '/^#/ ! {/\\$/ ! { /^[[:space:]]*$/ ! s/\([^ ]*\)\(.*\)$/\1\2 -J-Dgemfirexd.hostname-for-clients=\1/}}' "${SNAPPY_HOME_DIR}/conf/servers"
+sed -i '/^#/ ! {/\\$/ ! { /^[[:space:]]*$/ ! s/\([^ ]*\)\(.*\)$/\1\2 -hostname-for-clients=\1/}}' "${SNAPPY_HOME_DIR}/conf/servers"
 
 
 OTHER_LOCATORS=`cat locator_list | sed '1d'`
@@ -113,10 +127,7 @@ done
 sh "${SNAPPY_HOME_DIR}/sbin/snappy-start-all.sh"
 
 # Setup and launch zeppelin, if configured.
-if [[ "${ZEPPELIN_HOST}" != "zeppelin_server" ]]; then
-  if [[ "${ZEPPELIN_MODE}" = "NON-EMBEDDED" ]]; then
-    sh copy-dir.sh "${SNAPPY_HOME_DIR}" zeppelin_server
-  fi
+if [[ "${ZEPPELIN_HOST}" != "NONE" ]]; then
   for server in "$ZEPPELIN_HOST"; do
     ssh "$server" -o StrictHostKeyChecking=no "mkdir -p ~/snappydata"
     scp -q -o StrictHostKeyChecking=no ec2-variables.sh "${server}:~/snappydata"
